requirements.txt
pytest
requests
allure-pytest
faker
allure-python-commons
pydantic